+++ lib/flutter_sliding_toast.dart
@@ -1,4 +1,4 @@
-library flutter_sliding_toast;
+library;
 
 export 'package:flutter_sliding_toast/src/interactive_toast.dart';
 export 'package:flutter_sliding_toast/src/models/sliding_toast_setting.dart';